refactor(editor): Rename serializeNode to useWorkflowDocumentNodeSerialization
