+++ packages/frontend/editor-ui/src/app/stores/workflowDocument/useWorkflowDocumentNodeSerialization.ts
